// src/modules/lead/model/lead.status.ts
export enum LeadStatus {
  NEW = 'NEW',
  PROCESSING = 'PROCESSING',

  SCRAPING_PENDING = 'SCRAPING_PENDING',
  SCRAPED = 'SCRAPED',
  SCRAPING_SKIPPED = 'SCRAPING_SKIPPED',

  QUALIFICATION_PENDING = 'QUALIFICATION_PENDING',
  NOT_QUALIFIED = 'NOT_QUALIFIED',
  QUALIFIED = 'QUALIFIED',
  REJECTED = 'REJECTED',

  ANALYSIS_PENDING = 'ANALYSIS_PENDING',
  ANALYZED = 'ANALYZED',

  OUTREACH_PENDING = 'OUTREACH_PENDING',
  OUTREACHED = 'OUTREACHED',

  CONTACT_MISSING = 'CONTACT_MISSING',

  CONTACTED = 'CONTACTED',
  ENGAGED = 'ENGAGED',
  RESPONDED = 'RESPONDED',
  CONVERTED = 'CONVERTED',
}

// export enum LeadBusinessOutcome {
//   NONE = 'NONE',
//   CONTACTED = 'CONTACTED',
//   ENGAGED = 'ENGAGED',
//   RESPONDED = 'RESPONDED',
//   CONVERTED = 'CONVERTED',
// }